add swipe to return home for App

+++ src/types.h
@@ -4,13 +4,13 @@
 #include <Arduino.h>
 
 typedef struct{
-    uint16_t w;
-    uint16_t h;
+    int32_t w;
+    int32_t h;
 } Size;
 
 typedef struct{
-    uint16_t x;
-    uint16_t y;
+    int32_t x;
+    int32_t y;
 } Vector2;
 
 #endif
